Gridded data (S3 class "data.grid")

Defines data on a full regular (spatial) grid. Constructor function of the data.grid-class.

data.grid(
  ...,
  grid = NULL,
  window = NULL,
  mask = NULL,
  set.NA = FALSE,
  warn = FALSE
)

Arguments

...
vectors or arrays of data with length equal to prod(grid$n).

grid
a grid.par-class object (optional).

window
spatial window (values outside this window will be masked), currently an sp-object of class extending SpatialPolygons.

mask
logical; vector (or array) indicating the selected values (not masked).

set.NA
logical; If TRUE, the values corresponding to masked cells are set to NA.

warn
logical; If TRUE a warning message is generated when original data is masked.

Value

Returns an object of class data.grid, a list with the arguments as components.

Details

If parameter grid.par is not specified it is set from first argument.

S3 "version" of the SpatialGridDataFrame-class of the sp package.

See also