Fit a nonparametric geostatistical model

Fits a nonparametric (isotropic) geostatistical model (joint estimation of the trend and the variogram) by calling locpol, np.svariso.corr and fitsvar.sb.iso iteratively. At each iteration, the trend estimation bandwith is updated by a call to h.cv.

np.fitgeo(x, ...)

# S3 method for default
np.fitgeo(x, y, nbin = NULL, iter = 2, h = NULL,
  tol = 0.05, set.NA = FALSE, h.svar = NULL, maxlag = NULL,
  nlags = NULL, dk = 0, svm.resid = FALSE, warn = FALSE, plot = FALSE,
  ...)

# S3 method for locpol.bin
np.fitgeo(x, svm, iter = 1, tol = 0.05,
  h.svar = svm$esv$locpol$h, dk = 0, warn = FALSE, plot = FALSE, ...)

# S3 method for fitgeo
np.fitgeo(x, iter = 1, tol = 0.05,
  h.svar = x$svm$esv$locpol$h, dk = x$svm$par$dk, warn = FALSE,
  plot = FALSE, ...)

Arguments

x
object used to select a method (binned data, binned density or binned semivariogram).

...
further arguments passed to h.cv (trend bandwith selecction parameters).

y
vector of data (response variable).

nbin
vector with the number of bins on each dimension.

iter
maximum number of interations (of the whole algorithm). (traditional approach in geostatistics).

h
initial bandwidth matrix for trend estimation (final bandwith if iter <= 1).

tol
relative convergence tolerance (semivariogram).

set.NA
logical. If TRUE, sets the bin averages corresponding to cells without data to NA.

h.svar
bandwidth matrix for variogram estimation.

maxlag
maximum lag. Defaults to 55% of largest lag.

nlags
number of lags. Defaults to 101.

dk
dimension of the Shapiro-Botha variogram model (see fitsvar.sb.iso).

svm.resid
logical; if TRUE, the fitted (uncorrected) residual semivariogram model is computed and returned.

warn
logical; sets the handling of warning messages in bandwith selection (h.cv).

plot
logical; if TRUE, semivariogram estimates obtained at each iteration are plotted.

svm
(fitted) variogram model (object of class fitsvar or svarmod).

Value

Returns an object of class fitgeo (extends np.geo). A locpol.bin object with the additional (some optional) 3 components:

svm

fitted variogram model (object of class fitsvar).

svm0

(if requested) fitted residual variogram model (object of class fitsvar).

residuals

model residuals.

Details

Currently, only isotropic semivariogram estimation is supported.

If parameter h is not specified, h.cv is called to set it. If parameter h.svar is not specified, is set to 1.5*h.cv.svar.bin()$h.

See also

locpol, fitsvar.sb.iso, np.svar, np.svariso.corr, np.geo.