Interface to package "gstat"

Utilities to interact with the gstat package.

as.vgm(x, ...)

# S3 method for variomodel
as.vgm(x, ...)

# S3 method for svarmod
as.vgm(x, ...)

vgm.tab.svarmod(x, h = seq(0, x$range, length = 1000), sill = x$sill, ...)

# S3 method for sb.iso
as.vgm(x, h = seq(0, x$range, length = 1000), sill = x$sill, ...)

Arguments

x

variogram model object (used to select a method).

...

further arguments passed to or from other methods.

h

vector of lags at which the covariogram is evaluated.

sill

sill of the covariogram (or pseudo-sill).

Value

A variogramModel-class object of the gstat package.

Details

Tries to convert a variogram object to vgm (variogramModel-class of gstat package). S3 generic function.

as.vgm.variomodel tries to convert an object of class variomodel defined in geoR (interface to as.vgm.variomodel defined in gstat).

vgm.tab.svarmod converts a svarmod object to a variogramModel-class object of type "Tab" (one-dimensional covariance table).

as.vgm.sb.iso is an alias of vgm.tab.svarmod.

See also