npsp internal and secondary functions

Listed below are supporting functions for the major methods in npsp.

.cpu.time.ini()

revdim(a, d)

.compute.masked(bin, cov.bin = NULL, tol.mask = npsp.tolerance(2))

.wloss(est, teor, w, loss = c("MSE", "MRSE", "MAE", "MRAE"))

# S3 method for locpol.bin
residuals(object, ...)

.kriging.simple.solve(x, newx, svm)

# S3 method for np.geo
residuals(object, ...)

# S3 method for grid.par
print(x, ...)

# S3 method for grid.par
dim(x)

# S3 method for grid.par
dimnames(x)

# S3 method for grid.par
as.data.frame(x, row.names = dimnames(x), optional = FALSE, ...)

is.data.grid(x)

# S3 method for data.grid
dim(x)

# S3 method for data.grid
dimnames(x)

.rice.rule(x, a = 2, b = 3, ...)

splot.plt(
  horizontal = FALSE,
  legend.shrink = 0.9,
  legend.width = 1,
  legend.mar = ifelse(horizontal, 3.1, 5.1),
  bigplot = NULL,
  smallplot = NULL
)

.rev.colorRampPalette(colors, interpolate = "spline", ...)

Arguments

a

scale values.

cov.bin

(optional) covariance matrix of the binned data or semivariogram model (svarmod-class) of the (unbinned) data.

object

object used to select a method: local polynomial estimate of the trend (class locpol.bin) or nonparametric geostatistical model (class extending np.geo).

...

further arguments passed to or from other methods.

x

vector/matrix with data locations (each component/row is an observation location).

newx

vector/matrix with the (irregular) locations to predict (each component/row is a prediction location). or an object extending grid.par-class (data.grid).

svm

semivariogram model (of class extending svarmod).

b

exponent values.

Value

.compute.masked returns a list with components:

mask

logical vector bin$binw > tol.mask.

w

x$binw[mask].

sw

sum(w).

hat

(optional) bin$locpol$hat[mask, mask].

cov.bin

(optional) masked (aproximated) covariance matrix of the binned data.