Exploratory scatter plots

Draws (in a 2 by 2 layout) the following plots: a scatter plot with a color scale, the scatter plots of the response against the (first two) coordinates and the histogram of the response values.

scattersplot(x, ...)

# S3 method for default
scattersplot(x, z, main, xlab, ylab, zlab,
  col = hot.colors(128), lowess = TRUE, density = TRUE, omd = c(0.05,
  0.95, 0.01, 0.95), ...)

# S3 method for SpatialPointsDataFrame
scattersplot(x, data.ind = 1, main, xlab,
  ylab, zlab, col = hot.colors(128), lowess = TRUE, density = FALSE,
  omd = c(0.05, 0.95, 0.01, 0.95), ...)

Arguments

x
object used to select a method.

...
additional graphical parameters (to be passed to spoints).

z
vector of data (response variable).

main
an overall title for the plot.

xlab
a title for the axis corresponding to the first coordinate.

ylab
a title for the axis corresponding to the second coordinate.

zlab
a title for the axis corresponding to the response.

col
color table used to set up the color scale (see spoints).

lowess
logical. If TRUE, a lowess smooth is added to the plots of the response against the coordinates.

density
logical. If TRUE, a kernel density estimate is added to the histogram.

omd
a vector of the form c(x1, x2, y1, y2) giving the region inside outer margins in normalized device coordinates (i.e. fractions of the device region).

data.ind
integer (or character) with the index (or name) of the data component.

Details

Standard generic function with a default method, in which argument x is a matrix with the spatial coordinates (each row is a point). scattersplot.SpatialPointsDataFrame sets default values for some of the arguments from attributes of the object x (if present; see e.g. precipitation).

See also

splot, spoints, lowess, density