Linear binning of semivariances

Creates a svar.bin (binned semivar. + grid parameters) object with linearly binned semivariances (i.e. computes a binned sample variogram).

svar.bin(x, ...)

# S3 method for default
svar.bin(x, y, maxlag = NULL, nlags = NULL,
  minlag = maxlag/nlags, estimator = c("classical", "modulus"), ...)

svariso(x, y, maxlag = NULL, nlags = NULL, minlag = maxlag/nlags,
  estimator = c("classical", "modulus"), ...)

Arguments

x
object used to select a method. Usually a matrix with the coordinates of the data locations (columns correspond with dimensions and rows with data).

...
further arguments passed to or from other methods.

y
vector of data (response variable).

maxlag
maximum lag. Defaults to 55% of largest lag.

nlags
number of lags. Defaults to max(12, rule.svar(x)).

minlag
minimun lag.

estimator
character, estimator name (e.g. "classical"). See "Details" below.

Value

Returns an S3 object of class svar.bin (extends bin.data), a data.grid object with the following 4 components:

biny

array (dimension nlags) with the binned semivariances.

binw

array (dimension nlags) with the bin counts (weights).

grid

a grid.par-class object with the grid parameters.

data

a list with 3 components:

x argument x.

y argument y.

med (weighted) mean of the (binned) semivariances.

svar

a list of 2 components:

type character, type of estimation (e.g. "isotropic").

estimator character, estimator name (e.g. "classical").

Details

Currently, only isotropic semivariogram estimation is supported.

If parameter nlags is not specified is set to max(12, rule.svar(x)).

See also

np.svariso, np.svar, data.grid, binning, locpol, rule.svar.