Covariance matrix

Computes the covariance matrix a corresponding to a set of spatial locations given a variogram model or a semivariogram estimate.

varcov(x, coords, ...)

# S3 method for isotropic
varcov(
  x,
  coords,
  sill = x$sill,
  range.taper,
  discretize = nrow(coords) > 256,
  ...
)

# S3 method for np.svar
varcov(x, coords, sill = max(x$est), range.taper = x$grid$max, ...)

Arguments

x
variogram model (svarmod object) or semivariogram estimate.

coords
matrix of coordinates (columns correspond with dimensions and rows with data).

...
further arguments passed to or from other methods.

sill
(theoretical or estimated) variance \(C(0) = \sigma^2\) or pseudo-sill (unbounded variograms).

range.taper
(optional) if provided, covariances corresponding to distances larger than this value are set to 0.

discretize
logical. If TRUE (default), the variogram is (previously) discretized.

Value

The covariance matrix of the data.

See also