Plot a semivariogram object

Utilities for plotting pilot semivariograms or fitted models.

plot.fitsvar plots a fitted variogram model.

plot.svar.bin plots the binned semivariances.

plot.np.svar plots a local polynomial estimate of the semivariogram.

# S3 method for fitsvar
plot(
  x,
  y = NULL,
  legend = TRUE,
  xlab = "distance",
  ylab = "semivariance",
  xlim = NULL,
  ylim = c(0, 1.25 * max(x$fit$sv, na.rm = TRUE)),
  lwd = c(1, 2),
  add = FALSE,
  ...
)

# S3 method for svar.bin
plot(
  x,
  y = NULL,
  xlab = "distance",
  ylab = "semivariance",
  xlim = NULL,
  ylim = c(0, max(x$biny, na.rm = TRUE)),
  add = FALSE,
  ...
)

# S3 method for np.svar
plot(
  x,
  y = NULL,
  xlab = "distance",
  ylab = "semivariance",
  xlim = NULL,
  ylim = c(0, max(x$biny, na.rm = TRUE)),
  add = FALSE,
  ...
)

Arguments

x
a variogram object. Typically an output of functions np.svariso or fitsvar.sb.iso.

y
ignored argument.

legend
logical; if TRUE (default), a legend is added to the plot.

xlab
label for the x axis (defaults to "distance").

ylab
label for the y axis (defaults to "semivariance").

xlim
x-limits.

ylim
y-limits.

lwd
line widths for points (estimates) and lines (fitted model) respectively.

add
logical; if TRUE the semivariogram plot is just added to the existing plot.

...
additional graphical parameters (see par).

See also