Define a (semi)variogram model

Defines a variogram model specifying the parameter values. Constructor function of the svarmod-class.

svarmod(
  model,
  type = "isotropic",
  par,
  nugget = NULL,
  sill = NULL,
  range = NULL
)

svarmod.sb.iso(dk, x, z, nu, range, sill = nu)

svarmodels(type = "isotropic")

Arguments

model

string indicating the variogram family (see Details below).

type

string indicating the type of variogram, e.g. "isotropic".

par

vector of variogram parameters.

nugget

nugget value \(c_0\).

sill

variance \(\sigma^2\) or sill of the variogram (NA for unbounded variograms).

range

range (practical range or scale parameter) of the variogram (NA for unbounded variograms; maybe a vector for anisotropic variograms).

dk

dimension of the kappa function.

x

discretization nodes.

z

jumps (of the spectral distibution) at the discretization nodes.

nu

parameter \(\nu_0\) (can be thought of as the sill).

Value

svarmod returns an svarmod-class object, a list with function arguments as components.

svarmod.sb.iso returns an S3 object of class

sb.iso

(extends svarmod) corresponding to a `nonparametric' isotropic Shapiro-Botha model.

svarmodels returns a named character vector with the available models of the corresponding type

(when appropriate, component values could be used as cov.model argument in geoR routines and component names as model argument in gstat routines).

Note

svarmod does not check the consistency of the parameter values.

References

Shapiro, A. and Botha, J.D. (1991) Variogram fitting with a general class of conditionally non-negative definite functions. Computational Statistics and Data Analysis, 11, 87-96.

See also